- Using if condition in UserInformationActivity.kt in case of request already has been called; - Warzone request bug fixed (429 HTTP Error)

+++ app/src/main/java/com/example/callofdutymw_stats/viewmodel/UserInformationViewModel.kt
@@ -25,8 +25,9 @@ class UserInformationViewModel {
         fun responseKDRatioIsValid(response: String): Boolean {
             return response.toDouble() != 0.0
         }
+
         fun responseAccuracyIsValid(accuracy: String): Boolean {
-            return accuracy.toInt() != 0
+            return accuracy.toDouble() != 0.0
         }
     }
 }
